For Qt 6.8.+, make modules moved to extensions downloadable as modules
