+++ pyproject.toml
@@ -23,13 +23,14 @@ classifiers = [
 requires-python = ">=3.9"
 dependencies = [
     "bs4",  # canonical name is beautifulsoup4
-    "defusedxml",
-    "humanize",
-    "patch>=1.16",
+    "defusedxml~=0.8.0rc2",
+    "humanize~=4.11.0",
+    "patch~=1.16",
     "py7zr>=1.0.0-rc1",
-    "requests>=2.31.0",
-    "semantic-version",
-    "texttable",
+    "requests~=2.32.3",
+    "semantic-version~=2.10.0",
+    "texttable~=1.7.0",
+    "pyinstaller~=6.11.1"
 ]
 dynamic = ["version", "readme"]
 
@@ -130,7 +131,7 @@ disallow_incomplete_defs = false
 disallow_untyped_calls = false
 disallow_untyped_defs = false
 disallow_untyped_decorators = false
-# not a all 3rd party package has type hints
+# not an all 3rd party package has type hints
 ignore_missing_imports = true
 # Optional check: implicit check enabled in mypy 0.980 or before
 strict_optional = true
